- shut up compiler warnings
git-svn-id: svn://svn.paxle.net/paxle/trunk@1387 f5d3ab42-bacf-4b1a-9692-565be65a0cfe

+++ bundles/CryptUtils/src/main/java/org/paxle/crypt/impl/md5/MD5.java
@@ -181,42 +181,42 @@ public class MD5 {
 				| ((buffer[shift + 25] & 0xff) << 8)
 				| ((buffer[shift + 26] & 0xff) << 16)
 				| (buffer[shift + 27] << 24);
-		out[7] = ((int) (buffer[shift + 28] & 0xff))
-				| (((int) (buffer[shift + 29] & 0xff)) << 8)
-				| (((int) (buffer[shift + 30] & 0xff)) << 16)
-				| (((int) buffer[shift + 31]) << 24);
-		out[8] = ((int) (buffer[shift + 32] & 0xff))
-				| (((int) (buffer[shift + 33] & 0xff)) << 8)
-				| (((int) (buffer[shift + 34] & 0xff)) << 16)
-				| (((int) buffer[shift + 35]) << 24);
-		out[9] = ((int) (buffer[shift + 36] & 0xff))
-				| (((int) (buffer[shift + 37] & 0xff)) << 8)
-				| (((int) (buffer[shift + 38] & 0xff)) << 16)
-				| (((int) buffer[shift + 39]) << 24);
-		out[10] = ((int) (buffer[shift + 40] & 0xff))
-				| (((int) (buffer[shift + 41] & 0xff)) << 8)
-				| (((int) (buffer[shift + 42] & 0xff)) << 16)
-				| (((int) buffer[shift + 43]) << 24);
-		out[11] = ((int) (buffer[shift + 44] & 0xff))
-				| (((int) (buffer[shift + 45] & 0xff)) << 8)
-				| (((int) (buffer[shift + 46] & 0xff)) << 16)
-				| (((int) buffer[shift + 47]) << 24);
-		out[12] = ((int) (buffer[shift + 48] & 0xff))
-				| (((int) (buffer[shift + 49] & 0xff)) << 8)
-				| (((int) (buffer[shift + 50] & 0xff)) << 16)
-				| (((int) buffer[shift + 51]) << 24);
-		out[13] = ((int) (buffer[shift + 52] & 0xff))
-				| (((int) (buffer[shift + 53] & 0xff)) << 8)
-				| (((int) (buffer[shift + 54] & 0xff)) << 16)
-				| (((int) buffer[shift + 55]) << 24);
-		out[14] = ((int) (buffer[shift + 56] & 0xff))
-				| (((int) (buffer[shift + 57] & 0xff)) << 8)
-				| (((int) (buffer[shift + 58] & 0xff)) << 16)
-				| (((int) buffer[shift + 59]) << 24);
-		out[15] = ((int) (buffer[shift + 60] & 0xff))
-				| (((int) (buffer[shift + 61] & 0xff)) << 8)
-				| (((int) (buffer[shift + 62] & 0xff)) << 16)
-				| (((int) buffer[shift + 63]) << 24);
+		out[7] = (buffer[shift + 28] & 0xff)
+				| ((buffer[shift + 29] & 0xff) << 8)
+				| ((buffer[shift + 30] & 0xff) << 16)
+				| (buffer[shift + 31] << 24);
+		out[8] = (buffer[shift + 32] & 0xff)
+				| ((buffer[shift + 33] & 0xff) << 8)
+				| ((buffer[shift + 34] & 0xff) << 16)
+				| (buffer[shift + 35] << 24);
+		out[9] = (buffer[shift + 36] & 0xff)
+				| ((buffer[shift + 37] & 0xff) << 8)
+				| ((buffer[shift + 38] & 0xff) << 16)
+				| (buffer[shift + 39] << 24);
+		out[10] = (buffer[shift + 40] & 0xff)
+				| ((buffer[shift + 41] & 0xff) << 8)
+				| ((buffer[shift + 42] & 0xff) << 16)
+				| (buffer[shift + 43] << 24);
+		out[11] = (buffer[shift + 44] & 0xff)
+				| ((buffer[shift + 45] & 0xff) << 8)
+				| ((buffer[shift + 46] & 0xff) << 16)
+				| (buffer[shift + 47] << 24);
+		out[12] = (buffer[shift + 48] & 0xff)
+				| ((buffer[shift + 49] & 0xff) << 8)
+				| ((buffer[shift + 50] & 0xff) << 16)
+				| (buffer[shift + 51] << 24);
+		out[13] = (buffer[shift + 52] & 0xff)
+				| ((buffer[shift + 53] & 0xff) << 8)
+				| ((buffer[shift + 54] & 0xff) << 16)
+				| (buffer[shift + 55] << 24);
+		out[14] = (buffer[shift + 56] & 0xff)
+				| ((buffer[shift + 57] & 0xff) << 8)
+				| ((buffer[shift + 58] & 0xff) << 16)
+				| (buffer[shift + 59] << 24);
+		out[15] = (buffer[shift + 60] & 0xff)
+				| ((buffer[shift + 61] & 0xff) << 8)
+				| ((buffer[shift + 62] & 0xff) << 16)
+				| (buffer[shift + 63] << 24);
 	}
 
 	private native void Transform_native(int[] state, byte buffer[], int shift,
